==> notebooks/0_Data_Exploration.ipynb <==
---
title: "Chalice assessment: Data Exploration"
author: Liubov (Luba) Tovbin
date: "2024/11/19"
---

# Data Exploration

Here, we take an initial look at the data. Our goal is to predict whether or not a conversion will happen. Conversion prediction is a binary classification problem. We have two data files: one has the feature values that resulted in a conversion, and another has feature values that didn't result in a conversion. We want to know:

* How large is the dataset?
* How many data points do we have for each class?
* What features are there?
* What are the data types?
* Are there null values?
* What is the features' distribution?
* Are there outliers?

import pandas as pd
import seaborn as sns
sns.set_theme(context='notebook', style='whitegrid')
from helper_functions.descriptive_statistics import descriptive_statistics, generate_histograms, generate_html_report

# Load the data into pandas dataframe
conversion_df = pd.read_csv('../data/0_raw/Conversion_data.csv')
conversion_df.info()

nonconversion_df = pd.read_csv('../data/0_raw/nonconversion_data.csv')
nonconversion_df.info()

conversion_df.head()

nonconversion_df.head()

The conversion and non-conversion data are of the same format (except for the `SITE vs SITE column name difference). The conversion data frame has 100K rows, whereas the non-conversion data frame has 300K rows. Both datasets have fields with null values.
There are two numeric-valued columns, and the rest are categorical.

# Let's combine the conversion with nonconversion data in one data frame for convinience of exploration.
# Make columns' names the same
conversion_df.columns = nonconversion_df.columns

all_data = pd.concat([conversion_df, nonconversion_df]).reset_index(drop=True)
all_data.info()

all_data[:4]

## Numerical features' descriptive statistics

# Generate statistics
descriptive_statistics(
    all_data, 
    numerical_columns=all_data.select_dtypes(include=['number']).columns.to_list()
    )['descriptive_statistics']

A kurtosis of 9.069 for METRO values indicates that the distribution has heavy tails and a sharp peak, with significant extreme outliers. The high skewness of 94 suggests that the right tail is longer, while the majority of the data is concentrated on the left. The notable difference between the 75th quartile and the maximum value (554,053 - 669 = 553,384) further confirms this observation.

The standard deviation is substantial compared to the mean: 4,235 versus 657. Additionally, the interquartile range is narrow in relation to the overall data range: (669 - 526 = 143) compared to 554053. All of these factors reinforce the presence of extreme outliers.

In contrast, a kurtosis of -0.9 for USER_HOUR_OF_WEEK data indicates lighter tails and a flatter peak compared to a normal distribution. Here, the standard deviation is smaller than the mean, suggesting reasonable variability. The mean is less than 10% larger than the median, which confirms a light skewness of 0.16. The data range is 167, while the interquartile range is relatively narrow at (122 - 62 = 60). This indicates that 50% of the data falls within a moderately wide range of 60 units, suggesting that the data is fairly spread out.

Even though the METRO and USER_HOUR_OF_WEEK are presented as numerical features, they are categorical.

METRO values are the area codes; the difference between those numbers has no meaning. Thus, the values that appear as outliers might not necessarily be the ones. It would be beneficial to convert those area codes into coordinates to determine the data points' geographical distribution. We will keep the METRO values as is but treat them as categorical.

The same goes for the USER_HOUR_OF_WEEK feature. The values here range from 0 to 167, representing a consecutive hour in a 7-day week. That means an hour 0 is adjacent to the hour 167 of the previous week. We will treat this feature as categorical but transform it into more meaningful features: day and hour of the day.

# Let's have a look at numerical features distribution.
# After running this cell, go to output_images folder to see the plots
generate_histograms(
    all_data,
    numerical_columns=all_data.select_dtypes(include=['number']).columns.to_list()
)

## Categorical features' descriptive statistics

# Generate statistics
descriptive_statistics(
    all_data, 
    categorical_columns=all_data.select_dtypes(include=['object']).columns.to_list()
    )['descriptive_statistics']

The SITE column has 20687 unique values making it an extremely high cardinality feature. In addition, it seems that there are might be a lot of SITE values that appear only once or a small amount of times. The base 2 entropy of 10 indicates that on average, we'll need 10 bits to encode this feature. All of these suggest that the SITE feature might not be beneficial for our purpose.

## Histograms for low cardinality categorical features

Let's have a look at bar plots to get a better sense of categorical features distributions.

# Let's have a look at categrical features distribution.
# After running this cell, go to output_images folder to see the plots
generate_histograms(
    all_data,
    categorical_columns=all_data.drop(columns=['SITE']).select_dtypes(include=['object']).columns.to_list()
)

## Numerical features' statistics by category

# Generate statistics
descriptive_statistics(
    all_data,
    numerical_columns = ['METRO'],
    segment_by=['OS_FAMILY_NAME', 'BROWSER_NAME']
    )['descriptive_statistics']

generate_histograms(
    all_data,
    numerical_columns = ['METRO'],
    segment_by= ['OS_FAMILY_NAME', 'BROWSER_NAME']
)

## Report

Let's generate an HTML report of selected features.

# Define multiple segmented statistics functions
segmented_stats_func1 = lambda data: descriptive_statistics(
    data,
    numerical_columns=['METRO'],
    segment_by=['OS_FAMILY_NAME'],
    show_output=False
)['segmented_statistics']

segmented_stats_func2 = lambda data: descriptive_statistics(
    data,
    numerical_columns=['USER_HOUR_OF_WEEK'],
    segment_by=['OS_FAMILY_NAME', 'BROWSER_NAME'],
    show_output=False
)['segmented_statistics']

# Define multiple segmented histogram functions
segmented_histograms_func1 = lambda data, output_dir: generate_histograms(
    data,
    numerical_columns=['METRO'],
    segment_by=['OS_FAMILY_NAME'],
    output_dir=output_dir
)

#segmented_histograms_func2 = lambda data, output_dir: generate_histograms(
#    data,
#    numerical_columns=['USER_HOUR_OF_WEEK'],
#    segment_by=['OS_FAMILY_NAME', 'BROWSER_NAME'],
#    output_dir=output_dir
#)

# Generate HTML report
generate_html_report(
    all_data,
    numerical_stats_func=lambda data: descriptive_statistics(
        data,
        numerical_columns=data.select_dtypes(include=['number']).columns.to_list(),
        show_output=False
    )['descriptive_statistics'],
    categorical_stats_func=lambda data: descriptive_statistics(
        data,
        categorical_columns=data.select_dtypes(include=['object']).columns.to_list(),
        show_output=False
    )['descriptive_statistics'],
    segmented_stats_funcs=[segmented_stats_func1, segmented_stats_func2],
    numerical_histograms_func=lambda data, output_dir: generate_histograms(
        data,
        numerical_columns=data.select_dtypes(include=['number']).columns.to_list(),
        output_dir=output_dir
    ),
    categorical_histograms_func=lambda data, output_dir: generate_histograms(
        data,
        categorical_columns=data.drop(columns=['SITE']).select_dtypes(include=['object']).columns.to_list(),
        output_dir=output_dir
    ),
    segmented_histograms_funcs=[segmented_histograms_func1],
    output_dir="output_images",
    output_file="../report.html"
)
