==> notebooks/1_Data_Preparation.ipynb <==
---
title: "Chalice assessment: Data Preparation"
author: Liubov (Luba) Tovbin
date: "2024/11/19"
---

# Data Preparation

In this notebook, we create a clean, balanced training dataset and a testing dataset that emulates real-world data. In addition, we perform feature engineering to determine the optimal feature set and encode both datasets to get the model's input ready.

import pandas as pd
import seaborn as sns
sns.set_theme(context='notebook', style='whitegrid')
from sklearn.preprocessing import OrdinalEncoder

# Load the raw data
conversion_data = pd.read_csv('../data/0_raw/Conversion_data.csv')
nonconversion_data = pd.read_csv('../data/0_raw/nonconversion_data.csv')

## Data Cleaning

Let's create a clean, balanced training dataset by removing null values and combining the same number of conversion and non-conversion data points. The balanced dataset prevents the model from being biased towards the majority class (non-conversions).

We'll remove the null values for the testing dataset but keep class imbalance to emulate the real-world data.

# Clean column names by removing non-alphanumeric characters
conversion_data.columns = conversion_data.columns.str.replace(r'^[^a-zA-Z0-9]+|[^a-zA-Z0-9]+$', '', regex=True)
nonconversion_data.columns = nonconversion_data.columns.str.replace(r'^[^a-zA-Z0-9]+|[^a-zA-Z0-9]+$', '', regex=True)

# Check if ther columns' name are the same now
(conversion_data.columns == nonconversion_data.columns).all()

# Drop rows with null values
conversion_data = conversion_data.dropna().reset_index(drop=True)
nonconversion_data = nonconversion_data.dropna().reset_index(drop=True)

# Create a class column in the conversion and non-conversion datasets
# 1 means conversion 
# 0 means non conversion
conversion_data['CLASS'] = 1
nonconversion_data['CLASS'] = 0

# Shuffle the conversion data and perform a 80-20% train-test split 
conversion_data = conversion_data.sample(frac=1, random_state=11).reset_index(drop=True)
n = int(0.8 * len(conversion_data))
conversion_train = conversion_data[:n]
conversion_test = conversion_data[n:]

# Shuffle the non-conversion data and select n rows for training
nonconversion_data = nonconversion_data.sample(frac=1, random_state=11).reset_index(drop=True)
nonconversion_train = nonconversion_data[:n]
# For non-conversion test set, select three times more rows than in the conversion_test 
# to preserve the initial 3:1 ratio of non-conversion vs. conversion data.
m = len(conversion_test)
nonconversion_test = nonconversion_data[n: n + 3*m]

# Combine the conversion data non-conversion for train and test sets
train_data = pd.concat([conversion_train, nonconversion_train]).reset_index(drop=True)
test_data  = pd.concat([conversion_test, nonconversion_test]).reset_index(drop=True)

# Check if the data is balnced
train_data.groupby('CLASS').count()

The resulting training set is slightly imbalanced with more data points of class 1, the conversions. 
Since the XGBoost and Random Forest algorithms are robust to slight class imbalance, we won't rebalance the training set to avoid losing more data points.

# Save the clean data sets
train_data.to_csv('../data/1_processed/train_data.csv', index=False)
test_data.to_csv('../data/1_processed/test.csv', index=False)

## Feature Engineering (Light)

At first, we try only minimal feature manipulation to establish a baseline performance. Later, we will use more advanced feature selection techniques to see whether or not we can improve the model's performance.

We notice that the SITE column is problematic. Its high cardinality can lead to overfitting and lower model interpretability. We will drop the SITE column from the training set.

Lastly, we encode the categorical features. Since we are going to use tree-based methods for classification, we can use an OrdinalEncoder. Tree-based models decide splits based on thresholds rather than interpreting the encoded values as distances or magnitudes.

# Drop the SITE column from both, the train and the test sets
train_data = train_data.drop(columns=['SITE'])
test_data = test_data.drop(columns=['SITE'])

train_data[:3]

# Columns to encode
columns_to_encode = train_data.select_dtypes('object').columns

# Initialize the OrdinalEncoder with handling for unknown categories
ordinal_encoder = OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1)

# Fit the encoder on the training data
ordinal_encoder.fit(train_data[columns_to_encode])

# Transform both training and test data
train_data_encoded = train_data.copy()
test_data_encoded = test_data.copy()

train_data_encoded[columns_to_encode] = ordinal_encoder.transform(train_data[columns_to_encode])
test_data_encoded[columns_to_encode] = ordinal_encoder.transform(test_data[columns_to_encode])


test_data_encoded[:3]

# Save the encoded data sets
train_data_encoded.to_csv('../data/1_processed/train_data_encoded.csv', index=False)
test_data_encoded.to_csv('../data/1_processed/test_data_encoded.csv', index=False)
