==> notebooks/4_Model_Evaluation.ipynb <==
---
title: "Chalice assessment: Model Evaliation"
author: Liubov (Luba) Tovbin
date: "2024/11/19"
---

# Model Evaluation

import pandas as pd
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, roc_auc_score
from helper_functions.utils import evaluate_fairness_by_group

results_xgb = pd.read_csv('../data/3_evaluation/XGBClassifier_results.csv')
results_xgb_engineered = pd.read_csv('../data/3_evaluation/XGBClassifier_engineered_results.csv')
results_xgb_tunned = pd.read_csv('../data/3_evaluation/XGBClassifier_tunned_results.csv')

# Calculate  accuracy, global precision, recall, and F1-score
results = {
    'Model': ["XGB", "XGB_engineered", "XGB_tunned"],
    'Accuracy':  [],
    'Precision': [],
    'Recall':    [],
    'F1-Score':  [],
    'ROC-AUC':   [],
}

for df in [results_xgb, results_xgb_engineered, results_xgb_tunned]:
    results['Accuracy'].append(accuracy_score(df['CLASS'], df['CLASS_pred']))
    results['Precision'].append(precision_score(df['CLASS'], df['CLASS_pred']))
    results['Recall'].append(recall_score(df['CLASS'], df['CLASS_pred']))
    results['F1-Score'].append(f1_score(df['CLASS'], df['CLASS_pred']))
    results['ROC-AUC'].append(roc_auc_score(df['CLASS'], df['CLASS_pred']))

# Create a DataFrame
results_df = pd.DataFrame(results)

results_df

## Resulsts Discussion

The tuned XGBoost model with all initial features is the best-performing model, improving across all metrics. An attempt to modify a feature set resulted in a much lower performance across all metrics, even compared to the baseline, untuned XGBoost.

The XGBoost tuned correctly classified 83% of the instances. It is better at avoiding false positives, with 63% precision. Avoiding false positives is critical if we want to optimize the ad spend. The model best identifies actual conversions, with 81% recall, which is essential if missing a conversion is costly. 

The 71% F1 score confirms that the tuned model is better at balancing precision and recall. The 83% ROC-AUC shows the model's ability to distinguish between positive and negative classes across all thresholds, and it is also the highest for the tuned XGBoost model.


## Bias Detection

Let's check for biases in the model. We'll group the data by a potentially problematic feature, such as OS_FAMILY_NAME, and measure each group's **accuracy** and **selection rate**. 

The accuracy will give us a percentage of correct predictions within each group and the selection rate - a proportion of points in each group we predicted as conversions.

# Load the data before encoding
test_data = pd.read_csv('../data/1_processed/test.csv')

test_data[:3]

# Check that CLASS labels match
(results_xgb_tunned['CLASS'] == test_data['CLASS']).all()

X_test = test_data.drop(columns=['CLASS'])
y_test = test_data['CLASS']
y_pred = results_xgb_tunned['CLASS_pred'] 

sorted_metrics, overall_metrics = evaluate_fairness_by_group(
    y_true=y_test,
    y_pred=y_pred,
    sensitive_feature_column=X_test['OS_FAMILY_NAME'],
    metric="Accuracy"
)

print("Group-wise Metrics (Sorted by Accuracy):")
print(sorted_metrics)

print("\nOverall Metrics:")
print(overall_metrics)


sorted_metrics, overall_metrics = evaluate_fairness_by_group(
    y_true=y_test,
    y_pred=y_pred,
    sensitive_feature_column=X_test['BROWSER_NAME'],
    metric="Accuracy"
)

print("Group-wise Metrics (Sorted by Accuracy):")
print(sorted_metrics)

print("\nOverall Metrics:")
print(overall_metrics)


## Bias Analysis

The Windows users group has the highest selection rate and second to lowest accuracy. That indicates that our model is biased towards Windows users, predicting conversion for them more often than it should.

The Chrome users group has the highest selection rate with 3rd to lowest accuracy. However, we know that Chrome users dominate our dataset along with Windows users (obviously, the two largely intersecting users' sets). More data points mean more room for errors. (See output_images/OS_FAMILY_NAME_categorical.png and output_images/BROWSER_NAME_categorical.png)

The accuracy doesn't change dramatically across the OS_FAMILY_NAME or BROWSER_NAME groups, meaning the model is relatively fair.
